Removal of ICommunication Strategy and ROS2 alternative implementation to keep to VDA5050 explicit requirement for MQTT. Using 's MQTT Client for now
